Refactor pf-localization into a mrpt::apps class
Progress with #714
